Fixed bug by which font size was reset when changing color themes in PUCK. Fixes issue #304.

+++ puck/src/org/f2o/absurdum/puck/gui/panels/code/RSyntaxThemeUtil.java
@@ -24,12 +24,12 @@ public class RSyntaxThemeUtil
 		         ss = (SyntaxScheme) ss.clone();
 		         for (int i = 0; i < ss.getStyleCount(); i++) {
 		            if (ss.getStyle(i) != null) {
-		               ss.getStyle(i).font = ss.getStyle(i).font.deriveFont(theFontSize);
+		               ss.getStyle(i).font = ss.getStyle(i).font.deriveFont((float)theFontSize);
 		            }
 		         }
 		         ta.setSyntaxScheme(ss);
 		         
-		         ta.setFont(ta.getFont().deriveFont(theFontSize));
+		         ta.setFont(ta.getFont().deriveFont((float)theFontSize));
 			
 		}
 		else
